Refactor services to classes
