--- packages/js-toolkit/services/raf.ts
@@ -1,85 +1,0 @@
-/* eslint-disable no-use-before-define, @typescript-eslint/no-use-before-define */
-import { useService } from './service.js';
-import { useScheduler } from '../utils/scheduler.js';
-import { isFunction } from '../utils/is.js';
-import type { ServiceInterface } from './index.js';
-
-const scheduler = useScheduler(['update', 'render']);
-
-export type RafService = ServiceInterface<RafServiceProps>;
-
-export interface RafServiceProps {
-  time: DOMHighResTimeStamp;
-}
-
-/**
- * Create raf service.
- */
-function createRafService(): RafService {
-  let isTicking = false;
-
-  /**
-   * Trigger callbacks.
-   */
-  function trigger(props: RafServiceProps) {
-    for (const callback of callbacks.values()) {
-      scheduler.update(() => {
-        const render = callback(props) as (() => unknown) | void;
-        if (isFunction(render)) {
-          scheduler.render(() => {
-            // @ts-ignore
-            render(props);
-          });
-        }
-      });
-    }
-  }
-
-  /**
-   * Loop.
-   */
-  function loop() {
-    // eslint-disable-next-line no-use-before-define
-    props.time = performance.now();
-    trigger(props);
-
-    if (!isTicking) {
-      return;
-    }
-
-    requestAnimationFrame(loop);
-  }
-
-  const { add, remove, has, props, callbacks } = useService({
-    props: {
-      time: performance.now(),
-    } as RafServiceProps,
-    init() {
-      isTicking = true;
-      requestAnimationFrame(loop);
-    },
-    kill() {
-      isTicking = false;
-    },
-  });
-
-  return {
-    add,
-    remove,
-    has,
-    props: () => props,
-  };
-}
-
-let raf;
-
-/**
- * Use the RequestAnimationFrame (raf) service.
- */
-export default function useRaf(): RafService {
-  if (!raf) {
-    raf = createRafService();
-  }
-
-  return raf;
-}
